+++ resources/assets/js/account-import.js
@@ -1,0 +1,4 @@
+Vue.component(
+    'account-import',
+    require('./../components/AccountImport.vue').default
+);