+++ webpack.mix.js
@@ -34,6 +34,7 @@ mix.js('resources/assets/js/app.js', 'public/js')
 .js('resources/assets/js/spa.js', 'public/js')
 .js('resources/assets/js/stories.js', 'public/js')
 .js('resources/assets/js/portfolio.js', 'public/js')
+.js('resources/assets/js/account-import.js', 'public/js')
 .js('resources/assets/js/admin_invite.js', 'public/js')
 .js('resources/assets/js/landing.js', 'public/js')
 .vue({ version: 2 });